pyproject.toml
[project]
name = "music-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "opuslib>=3.0.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pynacl>=1.5.0",
    "pyyaml>=6.0.2",
    "yt-dlp>=2025.6.25",
]
